requirements.txt
ultralytics==8.3.159
langgraph==0.5.1
langchain==0.3.26
pyyaml==6.0.1
pillow~=11.3.0
langchain-core~=0.3.68
duckdb==1.3.2